requirements.txt
fastapi==0.115.5
python-dotenv==1.0.1
moviepy==2.1.1
yt-dlp==2024.11.18